moving dcraw code to where it belongs
