+++ public/filmstrip-worker.js
@@ -1,7 +1,41 @@
 /* global postMessage */
 
-const { imageUrl, imageUint8Array } = require('../renderer/tools/util.js');
+const fs = require('fs-extra');
+const dcraw = require('dcraw');
+
+const { bufferToUrl } = require('../renderer/tools/util.js');
 const log = require('../lib/log.js')('worker');
+
+async function readFileBuffer(filepath) {
+  if (Buffer.isBuffer(filepath)) {
+    return filepath;
+  }
+
+  const file = await fs.readFile(filepath);
+
+  return file;
+}
+
+async function imageUint8Array(filepath) {
+  log.time('preview');
+
+  const file = await readFileBuffer(filepath);
+
+  // read image from raw data
+  // var tiff = dcraw(file, { exportAsTiff: true });
+
+  const preview = dcraw(file, { extractThumbnail: true });
+
+  log.timeEnd('preview');
+
+  return preview;
+}
+
+async function imageUrl(filepath) {
+  const data = bufferToUrl(await imageUint8Array(filepath));
+
+  return data;
+}
 
 function exec(data) {
   function onDone(err, res) {
